Allow to run in peer-to-peer mode
Add --peer option that makes kio-fuse run internal D-Bus server which
is listening on a specified address instead of using the session D-Bus.

This makes it possible to start a separate instance of kio-fuse,
e.g. for testing. Also passing passwords in peer-to-peer mode
is more secure because messages can't be monitored unlike the
message bus.